Close Button Works

+++ src/main/java/ucf/assignments/Product.java
@@ -1,11 +1,13 @@
 package ucf.assignments;
 
+import java.math.BigDecimal;
+
 public class Product {
     private String name;
     private String serialNumber;
-    private Double value;
+    private BigDecimal value;
 
-    public Product(String name, String serialNumber, Double value) {
+    public Product(String name, String serialNumber, BigDecimal value) {
         this.name = name;
         this.serialNumber = serialNumber;
         this.value = value;
